+++ firmware/daisy/src/comm/daisy_inter_mcu_message_handlers.cpp
@@ -354,7 +354,9 @@ static void HandleSampleEditMessage(const uint8_t* payload, size_t payload_size)
                                       msg.start_frame,
                                       msg.end_frame,
                                       msg.loop_start,
-                                      msg.loop_end);
+                                      msg.loop_end,
+                                      msg.fade_in_ms,
+                                      msg.fade_out_ms);
 }
 
 static void HandleSampleMetaReqMessage(const uint8_t* payload, size_t payload_size) {